% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_regions.R
\name{add_regions}
\alias{add_regions}
\title{Add regional information to collision data}
\usage{
add_regions(data)
}
\arguments{
\item{data}{stays empty}
}
\value{
the reported collision dataset with additional columns containing information about the region
}
\description{
Add regional information to collision data
}
