% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_collision_types.R
\name{collision_types}
\alias{collision_types}
\title{Add collision types to collision data}
\usage{
collision_types(data, three = TRUE)
}
\arguments{
\item{data}{must be collision dataset of German Federal Statistical Office}

\item{three}{can be TRUE or FALSE, if TRUE more than two participating transport modes will be assigned to one value called "three", if FALSE more than two participating transport mdoes will be shown}
}
\value{
collision dataset with additional column "coll_typ" which contains information about which transport modes have collided
}
\description{
Add collision types to collision data
}
\examples{

}
