% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_2016.R
\name{import_2016}
\alias{import_2016}
\title{Import collisions with personal injury from year 2016}
\usage{
import_2016(x, save = FALSE)
}
\arguments{
\item{x}{stays empty}

\item{save}{if TRUE saves data frame as text file}
}
\value{
all collisions with personal injuries that have been reported in 2016
}
\description{
Import collisions with personal injury from year 2016
}
