% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_2017.R
\name{import_2017}
\alias{import_2017}
\title{Import collisions with personal injury from year 2017}
\usage{
import_2017(x, save = FALSE)
}
\arguments{
\item{x}{stays empty}

\item{save}{if TRUE saves data frame as text file}
}
\value{
all collisions with personal injuries that have been reported in 2017
}
\description{
Import collisions with personal injury from year 2017
}
