% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_2018.R
\name{import_2018}
\alias{import_2018}
\title{Title}
\usage{
import_2018(x)
}
\arguments{
\item{x}{stays empty}
}
\value{
all collisions with personal injuries that have been reported in 2018
}
\description{
Title
}
